rename the hook from useChat as we're no longer using a chat interface. and handle the direct json return without the answer
